refactor: backend structure to match ddd patterns
