--- packages/backend/src/infrastructure/messageProducer.ts
@@ -1,5 +1,0 @@
-import { Message } from '../messages/message'
-
-export interface MessageProducer {
-  send(message: Message): Promise<unknown>
-}
