+++ packages/backend/src/user/model.ts
@@ -1,4 +1,4 @@
-import { BaseModel } from './base'
+import { BaseModel } from '../shared/baseModel'
 
 export class User extends BaseModel {
   public static get tableName(): string {
